requirements.txt
wheel
python-registry